requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    bazel run //:pip_requirements.update
#
pyserial==3.5 \
    --hash=sha256:3c77e014170dfffbd816e6ffc205e9842efb10be9f58ec16d3e8675b4925cddb \
    --hash=sha256:c4451db6ba391ca6ca299fb3ec7bae67a5c55dde170964c7a14ceefec02f2cf0
    # via -r requirements.in
